# docs/FHIRAPI-swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "FHIR API",
    "description": "FHIR R4 and R5 view of the EHR. Sections show who calls each endpoint: payers during CRD, payer DTR apps, and the Payer Gateway (/internal).",
    "version": "v1"
  },
  "paths": {
...
    "/internal/access-tokens": {
      "post": {
        "tags": [
          "4. Internal - called by the Payer Gateway only"
        ],
        "summary": "Issues a short-lived access token for one patient, in the CDS Hooks \"fhirAuthorization\" shape.\nThe gateway sends it to the payer inside the CRD hook call, so the payer can read extra data that was not in the\nprefetch. The token is issued and kept (ITokenService), but no endpoint of this API checks it: authentication\nis added outside this code.",
        "requestBody": {
          "description": "Required. subject (payer) and patientId must be filled; scope and minutes are optional.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The token, ready to send as fhirAuthorization.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessTokenResponse"
                }
              }
            }
          },
          "400": {
            "description": "subject or patientId is missing.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthError"
                }
              }
            }
          }
        }
      }
    },
    "/internal/smart-launches": {
      "post": {
        "tags": [
          "4. Internal - called by the Payer Gateway only"
        ],
        "summary": "Creates a SMART launch context for a DTR app (patient, user, encounter, orders/coverage, appContext).\nThe gateway then redirects the browser to {DTR app}?iss={iss}&launch={launchId}.\n\"iss\" is the R4 base URL, or the R5 base URL when the request says fhirVersion = \"r5\".",
        "requestBody": {
          "description": "Required. patientId must be filled; user, encounter, fhirContext, appContext and fhirVersion are optional.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmartLaunchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The launch id and iss.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmartLaunchResponse"
                }
              }
            }
          },
          "400": {
            "description": "patientId is missing, or fhirVersion is not \"r4\" or \"r5\".",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OAuthError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AccessTokenRequest": {
        "required": [
          "patientId",
          "subject"
        ],
        "type": "object",
        "properties": {
          "subject": {
            "type": "string",
            "description": "Required. Who will use the token, normally the payer's id.",
            "example": "PAYER001"
          },
          "patientId": {
            "type": "string",
            "description": "Required. The only patient the token may read (EHR patient id).",
            "example": "pat-1"
          },
          "scope": {
            "type": "string",
            "description": "SMART scopes, separated by a space. Empty = read access to the patient's data (SmartScopes.PayerDefault).",
            "nullable": true,
            "example": "patient/Patient.rs patient/Coverage.rs patient/Observation.rs"
          },
          "minutes": {
            "type": "integer",
            "description": "Token lifetime in minutes, from 1 to 15 (a smaller or larger value is changed to 1 or 15).",
            "format": "int32",
            "default": 5,
            "example": 5
          }
        },
        "additionalProperties": false,
        "description": "Body of POST /internal/access-tokens (sent by the Payer Gateway)."
      },
      "AccessTokenResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "The token (a random value).",
            "example": "xhpE4XN-Z8MoI7ODwmf6JEQDnr6LFW60GmyUhVJCycw"
          },
          "token_type": {
            "type": "string",
            "description": "Always \"Bearer\".",
            "example": "Bearer"
          },
          "expires_in": {
            "type": "integer",
            "description": "Seconds until the token expires.",
            "format": "int32",
            "example": 300
          },
          "scope": {
            "type": "string",
            "description": "The scopes the token was given.",
            "example": "patient/Patient.rs patient/Coverage.rs patient/Observation.rs"
          },
          "subject": {
            "type": "string",
            "description": "Who the token was issued to (the payer id).",
            "example": "PAYER001"
          },
          "patient": {
            "type": "string",
            "description": "The patient the token is limited to.",
            "example": "pat-1"
          }
        },
        "additionalProperties": false,
        "description": "Response of POST /internal/access-tokens.\nSame shape as the CDS Hooks \"fhirAuthorization\" object, so the Payer Gateway can pass it to the payer unchanged."
      },
...
      "OAuthError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error code, e.g. invalid_request, invalid_grant, unsupported_grant_type.",
            "example": "invalid_request"
          },
          "error_description": {
            "type": "string",
            "description": "What went wrong, in plain text.",
            "nullable": true,
            "example": "subject and patientId are required."
          }
        },
        "additionalProperties": false,
        "description": "Standard OAuth 2.0 error body returned by the authorize, token and /internal endpoints."
      },
...
      "SmartLaunchRequest": {
        "required": [
          "patientId"
        ],
        "type": "object",
        "properties": {
          "patientId": {
            "type": "string",
            "description": "Required. Patient the DTR form is for (EHR patient id).",
            "example": "pat-1"
          },
          "userPractitionerId": {
            "type": "string",
            "description": "Logged-in clinician (becomes fhirUser = Practitioner/{id}).",
            "nullable": true,
            "example": "prac-1"
          },
          "encounterId": {
            "type": "string",
            "description": "Current encounter, if any.",
            "nullable": true,
            "example": "enc-1"
          },
          "fhirContext": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "References handed to the DTR app as fhirContext: the orders and the coverage the form is for.",
            "example": [
              "ServiceRequest/ord-1-hd",
              "Coverage/cov-1"
            ]
          },
          "appContext": {
            "type": "string",
            "description": "appContext from the payer's CRD card (which questionnaires to load, coverage-assertion-id). Passed to the app unchanged.",
            "nullable": true,
            "example": "{\"coverage-assertion-id\":\"702644806bf74c70aa65d2ca1a4a34b1\",\"questionnaire\":[\"http://example.org/fhir/Questionnaire/dialysis-incenter-hd\"]}"
          },
          "fhirVersion": {
            "type": "string",
            "description": "FHIR release the DTR app works with: \"r4\" (default) or \"r5\". Decides the \"iss\" returned, i.e. whether the app\nreads /fhir/r4 or /fhir/r5.",
            "nullable": true,
            "example": "r4"
          }
        },
        "additionalProperties": false,
        "description": "Body of POST /internal/smart-launches (sent by the Payer Gateway when a user opens a DTR link)."
      },
      "SmartLaunchResponse": {
        "type": "object",
        "properties": {
          "launchId": {
            "type": "string",
            "description": "Opaque launch id the DTR app sends back to /fhir/r4/auth/authorize.",
            "example": "6-jM9cmaBKMd4rFPAl-hQp_D"
          },
          "iss": {
            "type": "string",
            "description": "FHIR base URL of this API (the SMART \"iss\").",
            "example": "http://localhost:5100/fhir/r4"
          },
          "expiresAt": {
            "type": "string",
            "description": "The launch must be used before this time.",
            "format": "date-time",
            "example": "2026-10-07T18:50:40.2202026+00:00"
          }
        },
        "additionalProperties": false,
        "description": "Response of POST /internal/smart-launches.\nThe Payer Gateway redirects the browser to {dtrAppUrl}?iss={Iss}&launch={LaunchId}."
      },
...
    }
  },
...
}